+++ CHANGELOG.md
@@ -5,6 +5,10 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 
 ❤️ **Donate:** Enjoying MagicMirror²? [Please consider a donation!](https://magicmirror.builders/donate) With your help we can continue to improve the MagicMirror²
 
+### fixed
+
+- 2110, 2111, 2118 recurring full day events should not use timezone adjustment. just compare month/day
+
 ## [2.13.0] - Unreleased (Develop Branch - Please add your contributions to this release.)
 
 _This release is scheduled to be released on 2020-10-01._